+++ include/CableLoadConstraint.h
@@ -7,7 +7,7 @@
 #include <list>
 #include <string>
 
-#include "CableLoadCase.h"
+#include "include/CableLoadCase.h"
 
 /**
  * @par CONSTRAINT LIMIT TYPE OVERVIEW
