+++ include/Catenary.h
@@ -7,7 +7,7 @@
 #include <list>
 #include <string>
 
-#include "SupportLibrary.h"
+#include "include/SupportLibrary.h"
 
 /**
  * @par CATENARY OVERVIEW
