+++ src/Cable.cc
@@ -1,7 +1,7 @@
 // This is free and unencumbered software released into the public domain.
 // For more information, please refer to <http://unlicense.org/>
 
-#include "Cable.h"
+#include "include/Cable.h"
 
 CableComponent::CableComponent()
 {
