+++ src/SupportLibrary.cc
@@ -5,7 +5,7 @@
 #include <cmath>
 #include <string>
 
-#include "SupportLibrary.h"
+#include "include/SupportLibrary.h"
 
 double Round(const double& number, const int& precision)
 {
